+++ ruralmed/src/Doctors/images/greater-than-symbol.png
