+++ ruralmed/src/Doctors/images/less-than-symbol.png
