+++ ruralmed/src/components/PatientNavBar.css
@@ -1,0 +1,113 @@
+* {
+    margin: 0;
+    padding: 0;
+    font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif;
+}
+
+.containers {
+    display: flex;
+}
+
+.leftcontainer {
+    height: 100vh; /* Ensures the parent container takes full screen height */
+    width: 16%;
+    display: flex;
+}
+
+.navbar {
+    position: fixed;
+    height: 100vh;
+    width: 30vh;
+    background: #012f33;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: space-between;
+}
+
+.upper ul li, .lower ul li {
+    list-style: none;
+    border: 1px solid white;
+    border-radius: 10px;
+    color: #FFFFFF;
+    width: 20vh;
+    margin: 2vh 0;
+    padding: 1.5vh;
+    text-align: center;
+    cursor: pointer;
+    -webkit-border-radius: 10px;
+    -moz-border-radius: 10px;
+    -ms-border-radius: 10px;
+    -o-border-radius: 10px;
+}
+
+.upper ul li:hover, .lower ul li:hover {
+    background: #ffffff;
+    color: #000000;
+}
+
+.settings-menu {
+    position: relative;
+}
+
+.settings-dropdown {
+    display: none;
+    position: absolute;
+    top: -19vh; /* positions the dropdown below the settings item */
+    left: 0;
+    background-color: #012f33;
+    border-radius: 5px;
+    width: 20vh;
+    z-index: 1;
+}
+
+.settings-dropdown li {
+    list-style: none;
+    border: 1px solid white;
+    border-radius: 10px;
+    color: #FFFFFF;
+    margin: 2px 0;
+    padding: 1vh;
+    text-align: center;
+    cursor: pointer;
+}
+
+.settings-dropdown li:hover {
+    background: #ffffff;
+    color: #000000 !important;
+}
+
+/* Show the dropdown on hover */
+.settings-menu:hover .settings-dropdown {
+    display: block;
+}
+
+.logo p {
+    margin-top: 2vh;
+    font-size: 25px;
+    color: #65cccc;
+    font-weight: bold;
+    text-align: center;
+}
+
+/* Styling for selected menu item */
+/* Existing styles remain the same */
+
+/* Styling for selected menu item */
+.selected {
+    background: #FFFFFF;
+    color: #000000 !important; /* Enforces black font color when selected */
+}
+
+/* Styling for links */
+.upper ul li a,
+.lower ul li a {
+    text-decoration: none;
+    color: inherit; /* Inherit color from parent for active/hover states */
+}
+
+/* Change color of active link (while navigating) */
+.upper ul li a:active,
+.lower ul li a:active {
+    color: #000000;
+}
